git-svn-id: https://svn.cebitec.uni-bielefeld.de/svn/mgx@2164 2f99df6c-41e9-6383-b5e2-de53e728a4e2

+++ MGX-maven-ejb/src/main/java/de/cebitec/mgx/model/misc/NamedVector.java
@@ -8,9 +8,9 @@ package de.cebitec.mgx.model.misc;
 public class NamedVector {
 
     private final String name;
-    private final long[] data;
+    private final double[] data;
 
-    public NamedVector(final String name, final long[] data) {
+    public NamedVector(final String name, final double[] data) {
         this.name = name;
         this.data = data;
     }
@@ -19,7 +19,7 @@ public class NamedVector {
         return name;
     }
 
-    public long[] getData() {
+    public double[] getData() {
         return data;
     }
